requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
groq
torch>=2.1.0
sentence-transformers==2.5.1
faiss-cpu==1.8.0
pandas==2.2.1
python-dotenv==1.0.1
pydantic==2.6.3
numpy==1.26.4
requests==2.31.0
